処理を static class に分割
