+++ Assets/UniGLTF/Editor/MeshUtility/TabBoneMeshRemover.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4094f82d7ee108a4997d80a692116033
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
