+++ Assets/UniGLTF/Editor/MeshUtility/TabMeshIntegrator.cs
@@ -1,0 +1,61 @@
+using UniGLTF.M17N;
+using UnityEditor;
+using UnityEngine;
+
+namespace UniGLTF.MeshUtility
+{
+    public static class TabMeshIntegrator
+    {
+        public static bool OnGUI(GameObject _exportTarget)
+        {
+            var _isInvokeSuccess = false;
+            GUILayout.BeginVertical();
+            {
+                GUILayout.BeginHorizontal();
+                GUILayout.FlexibleSpace();
+                if (GUILayout.Button("Process", GUILayout.MinWidth(100)))
+                {
+                    _isInvokeSuccess = TabMeshIntegrator.Execute(_exportTarget);
+                }
+                GUILayout.EndHorizontal();
+            }
+            GUILayout.EndVertical();
+            return _isInvokeSuccess;
+        }
+
+        static bool Execute(GameObject _exportTarget)
+        {
+            if (_exportTarget == null)
+            {
+                EditorUtility.DisplayDialog("Failed", MeshProcessingMessages.NO_GAMEOBJECT_SELECTED.Msg(), "ok");
+                return false;
+            }
+            var go = _exportTarget;
+
+            var allComponents = go.GetComponents(typeof(Component));
+            var keyWord = "VRMMeta";
+
+            foreach (var component in allComponents)
+            {
+                if (component == null) continue;
+                var sourceString = component.ToString();
+                if (sourceString.Contains(keyWord))
+                {
+                    EditorUtility.DisplayDialog("Failed", MeshProcessingMessages.VRM_DETECTED.Msg(), "ok");
+                    return false;
+                }
+            }
+
+            if (go.GetComponentsInChildren<SkinnedMeshRenderer>().Length > 0 || go.GetComponentsInChildren<MeshFilter>().Length > 0)
+            {
+                MeshUtility.MeshIntegrator(go);
+                return true;
+            }
+            else
+            {
+                EditorUtility.DisplayDialog("Failed", MeshProcessingMessages.NO_MESH.Msg(), "ok");
+                return false;
+            }
+        }
+    }
+}
