+++ Assets/UniGLTF/Editor/MeshUtility/TabMeshIntegrator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 65a227dcf3cb5f34085bd6829894fb64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
