+++ Assets/UniGLTF/Editor/MeshUtility/TabMeshSeparator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8921d732b66ae134db97faa11eaa9b17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
