+++ Assets/UniGLTF/Editor/MeshUtility/TabStaticMeshIntegrator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9701e1f6aaed8d14283243a7a50be2df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
